requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
